FFT-MKL: switch back to MKL_LONG instead of MKL_INT64

+++ src/Numerics/Providers/Common/Mkl/SafeNativeMethods.cs
@@ -384,25 +384,25 @@ namespace MathNet.Numerics.Providers.Common.Mkl
         #region FFT
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long x_fft_free([In] ref IntPtr handle);
+        internal static extern int x_fft_free([In] ref IntPtr handle);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long z_fft_create([Out] out IntPtr handle, long n, double forward_scale, double backward_scale);
+        internal static extern int z_fft_create([Out] out IntPtr handle, int n, double forward_scale, double backward_scale);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long c_fft_create([Out] out IntPtr handle, long n, float forward_scale, float backward_scale);
+        internal static extern int c_fft_create([Out] out IntPtr handle, int n, float forward_scale, float backward_scale);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long z_fft_forward([In] IntPtr handle, [In, Out] Complex[] x);
+        internal static extern int z_fft_forward([In] IntPtr handle, [In, Out] Complex[] x);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long c_fft_forward([In] IntPtr handle, [In, Out] Complex32[] x);
+        internal static extern int c_fft_forward([In] IntPtr handle, [In, Out] Complex32[] x);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long z_fft_backward([In] IntPtr handle, [In, Out] Complex[] x);
+        internal static extern int z_fft_backward([In] IntPtr handle, [In, Out] Complex[] x);
 
         [DllImport(_DllName, ExactSpelling = true, SetLastError = false, CallingConvention = CallingConvention.Cdecl)]
-        internal static extern long c_fft_backward([In] IntPtr handle, [In, Out] Complex32[] x);
+        internal static extern int c_fft_backward([In] IntPtr handle, [In, Out] Complex32[] x);
 
         #endregion FFT
